refactor(WEG-130): Show the primary submit buton in both filter lists
