+++ src/components/WelcomeFilters.tsx
@@ -9,11 +9,7 @@ import { SecondaryButton } from './SecondaryButton'
 import { useRouter } from 'next/router'
 import { Phone } from './icons/Phone'
 import classNames from '@lib/classNames'
-import { PrimaryButton } from './PrimaryButton'
-import { useUrlState } from '@lib/UrlStateContext'
-import { useUserGeolocation } from '@lib/hooks/useUserGeolocation'
 import { TextLink } from './TextLink'
-import { useLabels } from '@lib/LabelsContext'
 
 export const WelcomeFilters: FC<{
   onGoBack: () => void
@@ -22,14 +18,7 @@ export const WelcomeFilters: FC<{
   const texts = useTexts()
   const { push } = useRouter()
   const isMobile = useIsMobile()
-  const [urlState] = useUrlState()
-  const labels = useLabels()
-  const { latitude, longitude } = useUserGeolocation()
-  const tags = urlState.tags || []
 
-  const filteredRecords = recordsWithOnlyLabels.filter((recordLabels) =>
-    tags.every((tagId) => recordLabels.find((labelId) => labelId === tagId))
-  )
   return (
     <>
       <div className="h-screen md:h-auto flex flex-col overflow-y-auto">
@@ -58,39 +47,6 @@ export const WelcomeFilters: FC<{
             {texts.welcomeFiltersText}
           </p>
           <FiltersList recordsWithOnlyLabels={recordsWithOnlyLabels} />
-          <PrimaryButton
-            className="md:max-w-sm"
-            onClick={() =>
-              void push({
-                pathname: '/map',
-                query: {
-                  ...urlState,
-                  ...(latitude && longitude ? { latitude, longitude } : {}),
-                },
-              })
-            }
-            disabled={tags.length > 0 && filteredRecords.length === 0}
-            tooltip={
-              tags.length > 0 && filteredRecords.length === 0
-                ? texts.filtersButtonTextFilteredNoResultsHint
-                : ''
-            }
-          >
-            {(tags.length === 0 || tags.length === labels.length) &&
-              texts.filtersButtonTextAllFilters}
-            {tags.length > 0 &&
-              filteredRecords.length === 1 &&
-              texts.filtersButtonTextFilteredSingular}
-            {tags.length > 0 &&
-              filteredRecords.length > 1 &&
-              texts.filtersButtonTextFilteredPlural.replace(
-                '#number',
-                `${filteredRecords.length}`
-              )}
-            {tags.length > 0 &&
-              filteredRecords.length === 0 &&
-              texts.filtersButtonTextFilteredNoResults}
-          </PrimaryButton>
           <div className="hidden md:block w-full mt-6">
             <TextLink href={texts.moreOffersKVBLinkUrl}>
               {texts.moreOffersKVBLinkText}
